Store extracted from index to reuse in test
